Divided one selectors to small pieces
