+++ src/components/apartment-detail-info/apartment-detail-info.jsx
@@ -6,10 +6,10 @@ import ReviewsList from "../reviews-list/reviews-list.jsx";
 import {Operation, ActionCreator} from "../../reducer/data/data.js";
 import {connect} from "react-redux";
 import Map from "../map/map.jsx";
-import {getOffersNearby} from "../selectors.js";
+import {getOffersNearby, getOpenOffer} from "../../reducer/data/selectors.js";
 // import ApartmentList from "../apartment-list/apartment-list.jsx";
 import ReviewsForm from "../reviews-form/reviews-form.jsx";
-import {getAutorisationStatus, getOpenOffer} from "../selectors.js";
+import {getAutorisationStatus} from "../../reducer/user/selectors.js";
 import {AuthorizationStatus} from "../../reducer/user/user.js";
 import BookmarkButton from "../bookmark-button/bookmark-button.jsx";
 import PageNotFound from "../page-not-found/page-not-forund.jsx";
