+++ src/components/app/app.jsx
@@ -7,7 +7,8 @@ import {Operation as OperationUser} from "../../reducer/user/user.js";
 import Main from "../main/main.jsx";
 import SingIng from "../sing-in/sing-in.jsx";
 import ApartmentDetailInfo from "../apartment-detail-info/apartment-detail-info.jsx";
-import {getOffers, getCheckedStatus} from "../selectors.js";
+import {getOffers} from "../../reducer/data/selectors.js";
+import {getCheckedStatus} from "../../reducer/user/selectors.js";
 import PrivateRoute from "../private-route/private-route.jsx";
 import Favorites from "../favorites/favorites.jsx";
 import LoadingPage from "../loading-page/loading-page.jsx";
